syncduke: headless door shim -- single-player Duke Nukem 3D over a BBS terminal
Replace Chocolate Duke3D's SDL video/input/timer back end with a headless shim
(syncduke_*.c/h) that captures the Build engine's 8-bit framebuffer and renders
it to the user's terminal via termgfx -- auto-selecting JPEG XL (on SyncTERM),
sixel, or the text/block fallback (so even a no-sixel terminal like conhost
works), with F4 to cycle tiers and DSR-ack auto-depth pacing.

Includes the 64-bit (LP64) port of the vendored engine (CON VM + renderer
pointer widening), DOOR32.SYS client-socket I/O, the SyncDOOM-style control
scheme + mouse steering + Controls Help, per-user save/config dirs (-home),
64-bit-safe save/load, the build script + installer + example config, and a
keymap unit test.

v1 is SINGLE-PLAYER ONLY (no multiplayer/lobby/audio yet), shareware Episode 1.
See README.md / SEAM.md.

+++ src/doors/syncduke/Engine/src/unix_compat.h
@@ -25,7 +25,14 @@
 
 // Horrible horrible macro: Watcom allowed memory pointer to be cast
 // to a 32bits integer. The code is unfortunately stuffed with this :( !
-#define FP_OFF(x) ((int32_t) (x))
+// SyncDuke 64-bit patch: widen to intptr_t so the pointer round-trip uses
+// ((void*)FP_OFF(ptr), e.g. initfastcolorlookup's colhere/colhead clear and
+// the palookup pointer builds) survive on LP64 instead of truncating to 32
+// bits and crashing. Sites that store the result into an int32_t global
+// (the 3D render inner loops: palookupoffs/asm3) still truncate -- those are
+// a deeper 32-bit assumption handled separately. No-op on a 32-bit build.
+#include <stdint.h>
+#define FP_OFF(x) ((intptr_t) (x))
 
 #ifndef max
 #define max(x, y)  (((x) > (y)) ? (x) : (y))
